Implemented "texture-pack" region flag.
This implements a "texture-pack" region flag which will switch
textures in the Minecraft client as the user moves into a region
with this flag defined.

The current implementation adds a Maven dependency to CraftBukkit to
access the base Minecraft and CraftBukkit server classes, because
CraftPlayer.sendPluginMessage() does not support sending to the
internal Minecraft channels.
For details see: http://bukkit.atlassian.net/browse/BUKKIT-2579