Added experimental custom flag support
The custom flags work as follows:

1. WorldGuard's onEnable() is called.

2. World Guard looks for any plugin that implements CustomFlagProvider

3. World Guard invokes addCustomFlags() on each found plugin

4. World Guard passes in a CustomFlagBroker.

5. The plugin calles CustmFlagBroker.addFlag to add any flags it needs

A plugins addFlag method will be called afeter onLoad(), but may be called
before onEnable().

If a plugin that adds custom flags failes to load or is removed, World Guard
will remove all of the custom flags from all regions.
(Perhaps this can be changed in the future)